+++ test/components/gcs-map/GCSMap.spec.js
@@ -4,6 +4,7 @@ enableFetchMocks();
 import GCSMap from "../../../src/components/gcs-map/GCSMap";
 import * as defaultConfig from "../../../src/components/gcs-map/assets/config.json";
 import * as customConfig from "./assets/config.json";
+import * as customConfig2 from "./assets/config2.json";
 
 describe("Init gcs-map", () => {
 
@@ -263,5 +264,29 @@ describe("Reading of config.json", () => {
 
         expect(component.layerManager.backgroundLayers.length).toBe(2);
         expect(component.layerManager.foregroundLayer).toBeNull();
+    });
+
+    it("should initialize StyleManager correctly with style defined", async () => {
+        fetch.mockResponseOnce(JSON.stringify(customConfig2));
+        const component = new GCSMap();
+
+        console.error = jest.fn();
+
+        component.setAttribute("config-url", "http://config.service/config.json");
+
+        await component.connectedCallback();
+
+        expect(Array.isArray(component.styleManager.styleList)).toBe(true);
+        expect(component.styleManager.styleList.length).toBe(1);
+        expect(console.error.mock.calls.length).toBe(0);
+    });
+
+    it("should initialize StyleManager correctly without style defined", async () => {
+        const component = new GCSMap();
+
+        await component.connectedCallback();
+
+        expect(Array.isArray(component.styleManager.styleList)).toBe(true);
+        expect(component.styleManager.styleList.length).toBe(0);
     });
 });